mix small issues

+++ src/daemon/sockets.cpp
@@ -80,7 +80,7 @@ int scks_cleanup() {
 
     std::vector<int> todel;
 
-    for (int idx = 0; idx < gdata.sockets.size(); idx++) {
+    for (size_t idx = 0; idx < gdata.sockets.size(); idx++) {
         SocketInfo& sock_info = gdata.sockets[idx];
         // filter sockets that are idle for 15 seconds or more
         if (sock_info.fd == -1 || curr_time - sock_info.last_seen_ms < 15'000)
